Randomized World Generation
Implement code that takes in a seed, generates a pseudo-random 2D world array, and displays that world back to the user.

For Project 3A, when the user runs the main method in Main.java, they should see a world pop up. For now, you can hard-code the seed in the main method (and change the code to see the output of different seeds).

Besides the Main.main requirement, your world generation code can live anywhere in the repo (we suggest making your own files in the Core package).